SGI 64 bit warnings rampage

+++ CCA/Components/MPM/PhysicalBC/LoadCurve.h
@@ -1,7 +1,9 @@
 #ifndef UINTAH_MPM_LOADCURVE_H
 #define UINTAH_MPM_LOADCURVE_H
 
+#include <sgi_stl_warnings_off.h>
 #include <vector>
+#include <sgi_stl_warnings_on.h>
 #include <Packages/Uintah/CCA/Components/MPM/PhysicalBC/MPMPhysicalBC.h>
 #include <Core/Geometry/Vector.h>
 #include <Core/Geometry/Point.h>
@@ -71,7 +73,7 @@ WARNING
       // Get the load at time t
       inline T getLoad(double t) {
 
-         int ntimes = d_time.size();
+	int ntimes = static_cast<int>(d_time.size());
          if (t >= d_time[ntimes-1]) return d_load[ntimes-1];
 
          for (int ii = 1; ii < ntimes; ++ii) {
